pyproject.toml
[build-system]
requires = [
    "setuptools>=68",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "vla-scratch"
version = "0.1.0"
description = "Utilities for working with VLA Scratch datasets and policies."
requires-python = "==3.10.*"
dependencies = [
    "torch",
    "torchvision",
    "triton",
    "numpy",
    "jaxtyping",
    "tqdm",
    "einops",
    "tensordict",
    "transformers==4.57.1",
    "lerobot @ git+https://github.com/huggingface/lerobot@v0.4.2",
    "num2words", # for smolvlm
    "jsonlines",
    "pillow",
    "torchvision==0.22.0",
    "hydra-core",
    "omegaconf",
    "wandb",
    "setproctitle",
    "setuptools",

    # serve policy
    "pyzmq",
    "msgpack",
    "matplotlib",
    # the last commit before dataset version 3.0
    "lerobot @ git+https://github.com/huggingface/lerobot@v0.4.2",
    "av",  # pyav for video decoding (alternative to torchcodec)
    "accelerate",
    "mujoco",
    "art",
    "emoji>=2.15.0",
    
    # debug
    "pdbpp",

    # NOT USED
    # "matplotlib",
    # "mujoco",
    # "h5py",
    # "accelerate",
    # "viser",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "vla_scratch",
    "vla_scratch.*",
]
namespaces = true

[tool.uv]
required-environments = ["sys_platform == 'linux' and platform_machine == 'x86_64'"]
index-strategy = "unsafe-best-match"
override-dependencies = ["torch", "torchvision"]

[[tool.uv.index]]
name = "pytorch-nightly-cu128"
url = "https://download.pytorch.org/whl/nightly/cu128"
default = true

[[tool.uv.index]]
url = "https://pypi.org/simple"

[tool.uv.sources]
torch = { index = "pytorch-nightly-cu128" }
torchvision = { index = "pytorch-nightly-cu128" }
triton = { index = "pytorch-nightly-cu128" }

# python-install-mirror = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
